+++ sbncode/EventGenerator/MeVPrtl/Tools/RayTraceBox_tool.cc
@@ -14,7 +14,7 @@
 
 // local includes
 #include "IRayTrace.h"
-#include "../Products/MeVPrtlFlux.h"
+#include "sbnobj/Common/EventGen/MeVPrtl/MeVPrtlFlux.h"
 
 // LArSoft includes
 #include "larcorealg/Geometry/BoxBoundedGeo.h"
@@ -53,7 +53,7 @@ public:
     bool IntersectDetector(MeVPrtlFlux &flux, std::array<TVector3, 2> &intersection, double &weight) override;
 
     // no weights
-    float MaxWeight() override { return -1.; }
+    double MaxWeight() override { return -1.; }
 
 private:
   geo::BoxBoundedGeo fBox;
