+++ sbncode/EventGenerator/MeVPrtl/Tools/WeightedRayTraceBox_tool.cc
@@ -14,7 +14,7 @@
 
 // local includes
 #include "IRayTrace.h"
-#include "../Products/MeVPrtlFlux.h"
+#include "sbnobj/Common/EventGen/MeVPrtl/MeVPrtlFlux.h"
 #include "sbncode/EventGenerator/MeVPrtl/Tools/Constants.h"
 
 // LArSoft includes
@@ -60,7 +60,7 @@ public:
     bool IntersectDetector(MeVPrtlFlux &flux, std::array<TVector3, 2> &intersection, double &weight) override;
 
     // always thrown at least once
-    float MaxWeight() override { 
+    double MaxWeight() override { 
       return fMaxWeight;
     }
